combine juju-artifacts into 1 file
